web_renderjs_ui_test: fix flaky testFormViewHistoricalRevisionLink
This test was using a blind `pause` instead of an explicit waitFor*, so
it was failing when machine was not fast enough.

Also refactor the test a bit to simplify the assertion by first
selecting the iframe we are going to inspect and make the test
compatible with both chrome and firefox - they behave differently when
interpreting new lines in selenium test tables.