web_renderjs_ui_test: fix flaky testFormViewHistoricalRevisionLink
This test was using a blind `pause` instead of an explicit waitFor*, so
it was failing when machine was not fast enough.

Also refactor the test a bit to simplify the assertion by first
selecting the iframe we are going to inspect.

One strange thing is that the diff on that line_list property is
supposed to be None -> ('Foo', 'Bar') but this seem to be represented as
('Foo' 'Bar',) when running on testnodes ... for now we support both
patterns