dms_conversion_catalog: use proper activity queue
Because this is using SQLDict, when we had multiple conversion
activities for the same path with different conversion parameters, only
one was executed.

We don't want this here, as the recent failures on
testDmsWithPreConversion.TestDocumentWithPreConversion are showing, when
the same image is referenced with different conversion parameters we
want to pre-convert for all conversion parameters.

Use SQLQueue, which allows calling the same method more than once.