...
This is a follow up commit to cd18daf41128ac5ba69dafd62fbc72b43103f680,
to fix the behavior in WSGI.

Zope WSGI no longer works like Medusa in the case of "response.write".
Before this change, the response contained the contents of
"response.stdout" (which we want), plus the contents of the body (the
traceback created by the raised Exception, which we do not want). Note
that the "content-length" header was not correctly set either, and parts
of the response were cut off.

Reminder: raising an exception is done in order to doom the transaction.

We are working on reimplementing WSGI/Medusa compatible streaming in a
better way in the near future; this commit is a temporary fix.