ERP5Type: add a transaction note when bootstraping
Bootstrap is actually the automatic migration of tools. It happened to me
several times that I mistakenly start an old instance on a too new version of
products and this instance is migrated automatically.

By leaving a note in transaction it's easier to find out which
transaction performed the automatic migration to revert to the state
before using technique such as
https://blog.gocept.com/2019/06/14/undo-transactions-by-truncating-zodb-data-fs/
for FileStorage.

with this change, a command such as;

    python -m ZODB.scripts.fstail -n 20 ./instance/slappart6/srv/zodb/root.fs

show something like:

    2019-11-12 02:46:42.106313: hash=44fecb9a63411fbf305cea82f4c699c478000d61
    user=' None' description='/timer_service/process_timer\nSite migrated' length=462938 offset=881877846 (+70)

for the transaction which did the migration, so it's easy to repair.