hal_json_style/core: allow to set portal_status_message during edit
When passing workflow transitions, workflow scripts can change the
default status message by setting request key "portal_status_message" to
a specific custom message.

Extend this to Base_edit, so that it's also possible to set a custom
message for editions. In ERP5JS mode, portal_status_level is also
supported.

Same as with workflow scripts, "portal_status_message" must be set to an
already translated message.

This can be used for example with interaction workflows.