core: re-validate component after modify after transition
the "end of transaction" mode was not compatible with ERP5JS, because
form is re-rendered after edition in the same transaction.

Because this interaction can modify the workflow state and the source
code warnings/errors properties, we want this interaction to be executed
in the current transaction so that we can render the edited document
properly.

Also unset "once per transaction" for the case the same document is
edited more than once in the same transaction.
With this change, if that scenario of a document edited more than once
in a transaction really happen, then the source code of the document
will be checked multiple times.