preferences: use an interaction workflow to clear cache
When we use preference.edit(preferred_something=something_else) the
cache was cleared, because there is an interaction on _edit, but when we
do preference.setPreferredSomething(something_else) cache was not
cleared.

This still have a problem cache is cleared too much but at least it
eliminates this incorrect behavior.

Update tests at the same time:
 - testEditorField,testAuthenticationPolicy: no need to clear cache,
   it's now done automatically.
 - testUpgrader: the workflow chain of preference changed. This test is
   asserting the actual workflow chain, so we have to update the test
   everytime workflow chain is modified.