preferences: use an interaction workflow to clear cache
When we use preference.edit(preferred_something=something_else) the
cache was cleared, because there is an interaction on _edit, but when we
do preference.setPreferredSomething(something_else) cache was not
cleared.

This still have a problem that this is clearing cache too much but at
least it eliminates this incorrect behavior.

Update tests at teh same time:
 - testEditorField: no need to clear cache, it's now done automatically.
 - testUpgrader: the workflow chain of preference changed.