forge/administration: move some scripts to administration
These scripts are used to check the site configuration and also in
CodingStyleTest.

This is just the minimal so that CodingStyleTest can run, but this could
be extended to several scripts currently in erp5_toolbox skinfolder of
erp5_forge which overlaps with erp5_administration. Also, regarding
CodingStyleTest, it seems that erp5_administration contain a lot of
things, so we may want to extract a smaller business template containing
only the scripts used for static analysis of business template.