testXHTML: test portal types also from tools
Some portal types such as Business Template were not tested by testXHTML
because this tests starts by modules and recursively tests based on
allowed content types. But types that are not created in a module but in
a tool were never tested.

Keep the current way of testing view of portal types in their natural
containers by starting from modules and make another pass to test anyway
portal types that might be contained in tool. In that case, we use a
mapping of known tools IDs because I don't think we can introspect at
this point.