core: rename PropertySheetChild_validateIDField
PropertySheetChild is not an acceptable prefix, use Base as the context
is used. Technically, this uses the "id as reference" mixin, but this
mixin is a bit special as it has takes the suffix as parameter and does
not reside in mixin module, so we just use Base as a prefix.