patches: patch OFS to bring back old __repr__
When updating persistent to >= 4.4 __repr__ of zope objects display
information about ZODB connection, but it's more useful to have
information about the acqusition chain at this level.

One notable case was proxy field error messages, it was supposed to
include the path of the proxy field in the error message. This one is
fixed by fixing the MRO so that persistent.Persistent does not appear
first.