ingestion: split erp5_ingestion in erp5_ingestion_test
erp5_ingestion contained a skin folder for tests, with some proxy fields depending on erp5_web.
There was a proper test dependency to erp5_web, but in static analysis check we now check the skin folders, so it's no longer allowed to have some skin folder that are broken without test dependencies but not supposed to be used outside of tests.