core/forge: fix scripts "Repository History" actions
The action was in erp5_core, but the form in erp5_forge. Move the action
in erp5_forge.

The listbox list method was raising error because it was using catalog
with unsupported installation_state=. Change to use contentValues
instead.

This listbox had "Listbox" as a title, use "History" instead, which
makes a little more sense in this context.