VideoField: make valid HTML5 regarding preload
Unlike other attributes managed in video field, preload [1] is not a
boolean attribute, so setting preload="preload" was invalid HTML5.

On the field, we can only configure video_preload as a boolean
attribute, setting preload="auto" when true and not setting preload when
False seems the closest meaning of True/False.

[1] https://html.spec.whatwg.org/multipage/media.html#attr-media-preload