core: Workaround invalid html on AlarmTool_viewAlarmList
When listboxs has an editable field, but the field is non editable, it
produces a markup like:
  <a href="link to the line"><EditableField/></a>
This is fine for most fields, but some editable fields are rendered as
an <input> even if they are not editable - this is the case for
CheckBoxField.
To prevent rendering <a><input></a> which is not valid HTML, configure
the enabled field as non-editable on alarm list view, with this change
listbox treats the field as an editable field and just render the field
as <EditableField/> without the <a>.
To keep the same visual appearance of having a disabled checkbox field,
make this field disabled with extra.