xhtml_style: minimal emulation of router "raw" command
In xhtml style, clicking on <a> links in a non editable ck editor does nothing. It's possible to to use Ctrl+right click to trigger the popup menu and "open in new tab", but that's confusing for users.

Implement something minimal so that the event handler registered in ck editor gadget also works in html style.