*: stop using newTemp* constructors
These constructors are not recognized by pylint and now we have
no-name-in-module errors.

Because since introduction of portal types as classes there was this
idea of deprecating those in favor of using newContent(temp_object=True,
switch to this new syntax.

Since we still have to create documents in portal_trash, only change the
minimal to get coding style tests passing again.