ERP5Type.Core.Folder.Folder: Inherit from our classes first.
Allows Base to consistently override methods which are also overridden in
CMF where CMF implementation does not propagate the call to other
superclasses (ex: __before_publishing_traverse__ as of CMFCore-2.2.10).
Also allows simplifying several inheritance fixups.
Also, use super() instead of explicit resolution to Base.

Also: testBusinessTemplate: drop dead monkey-patch.
_getCopy is called during BusinessTemplate installation, but not the one
on this class. Which prevents reordering class inheritance between
SimpleItem and copy support classes.
So this is dead code which is not even explaining what it is trying to do
(ex: what document should it really not be called on ? what effect of this
method is undesired ? why is the method attached to this class ? why is
the cleanup restoring the original method, when this class does not have
this method to begin with but it merely gets it from its superclasses
which are re-inherited from by documents hence overriding it anyway ?,
and it is getting in the way of reordering class inheritance so it goes
away.

Also: ERP5Type.patches.CMFBTreeFolder: Fold patch onto ERP5Type.Core.Folder.
All ERP5UI-visible (hence needing to call allowedContentTypes) BTReeFolder2
instances should already inherit ERP5Type.Core.Folder, removing the need
for this monkey-patch.