Revert "erp5_dms_ui_test:  Add selenium tests"
This reverts commit f93918473edffb98ec3db36447d81d4f7f27884c.

These two paths were only half-committed, for the PDF we don't have the
file content, for the test we don't have the .xml, so in practice this
test never ran.
I tried to run it but it does not pass, uses pause command and some
xpath selector that are now unclear.

PDFJS is also tested in portal_tests/erp5_km_zuite/testCheckPDFPreview