ZODB Components: erp5_item: Migrate Item Document from FS and its Unit Test.
Deleted Device Document as it is not used anywhere and not really useful:
Item can be used instead.

Migrated SubscriptionItem from FS to erp5_open_trade: this is the contract as
in https://www.erp5.com/erp5-TechnicalNote.Open.Order#contract and make
erp5_payroll depends on erp5_open_trade as Employment Contract is a
SubscriptionItem and, per design, Payroll is kind of an Open Order.