Folder: fix checkConsistency on folder containing non erp5 objects
contentValues() nowadays return also non ERP5-ish objects such as python
scripts , OFS.File etc. Because of that, if an ERP5 Folder contained a
non ERP5 object, which does not have fixConsistency or checkConsistency
methods, Folder.checkConsistency was causing an infinite loop because it
was acquiring checkConsistency method from self.