erp5_web_renderjs_ui: Fixup in obtaining the translation source file list
previous version would keep only one file ending in 'translation_data.js'
which was buggy. The scope here is to return all files,
but put those ending with 'translation_data.js' first,
so that it is more user friendly for user following the convention.

Again for user friendliness, we sort the lists alphabetically also