core: whitespace preserve security_uid_column field
Since catalog is migrated to an ERP5 document and can only be edited
with ERP5 UI we lost the hability to define security uid column for
default group, which is supposed to be defined with an empty string,
something like: `  | security_uid`.

Enable whitespace preserve on this field to allow saving this string
string without being stripped by formulator.