scalability_data: remove
This business template is not used and contain a preference referencing
non existing paths, which causes errors in coding style test with
test_moduleListMethod that cannot display accounting module, because
preference is set to group/my_group which does not exist.

Instead of trying to fix, let's remove the business template if it's not
used.