core/administration: move BusinessTemplate_getPythonSourceCodeMessageList to core
This script is used in coding style tests, but because it was in
erp5_administration, all coding style tests had to install
erp5_administration and all its dependencies. Which is against one idea
of the coding style test, which is supposed to check that business
templates declare their dependencies correctly, but if we install all
business templates that erp5_administration depends on, the test might
have false positives of missing dependencies.