trade: move First Causality Movement Group allowed content type
This portal type is defined in erp5_trade, and the allowed content type in
Delivery Builder is defined in erp5_trade, only the allowed content type we
defined in erp5_base, which is incorrect, because it makes erp5_trade
allow some non existing types in Order Builder.

Fix this by moving the allowed content type in erp5_trade.