newTemp*() deprecation (1bce856, 04b4985): Remove hack in NotificationTool.sendMessage() not working with ZODB Components.
When a Portal Type could not be found, sendMessage() was creating a temp_object
instead (even when passing store_as_event=True) by calling newTempPORTAL_TYPE()
and thus assuming a filesystem Document. So from now on, Portal Type must be
available and thus sendMessage() will fail otherwise. Moreover, if store_as_event
is True, this will no longer create a temp_object silently as it used to when
the Portal Type is not available.

This moves Mail Message Portal Type from erp5_crm to erp5_base as MailMessage_send
(send() being called directly from sendMessage() for temp_object) is already
in erp5_base. As this is only to allow creating temp_object, leave its Actions,
Workflows and Constraint relying on erp5_crm API as they are.