ERP5Type: Don't log when cache factory is not found
This should prevent all the warnings like

    WARNING Cache.__call__ Factory erp5_content_medium not found, method <function getTypeList at 0x7fd8b2ce4cd0> executed without cache
    WARNING Cache.__call__ Factory erp5_content_long not found, method <function _getPortalGroupedTypeSet at 0x7fd8bc613750> executed without cache

we have during startup.

This code tried to detect the case where cache system was not
fully initialized an only log in this case, but there does not seem to
be a reliable way to detect that cache is supposed to be ready and that
log was not really useful - If some methods are executed without cache
for some reason (which as far as I know never happens), then we should
be able to notice the problem because it's slow.