ERP5TypeTestSuite: let testsuite decide the log directory
Using the test class name was causing conflicts on some test classes.
CodingStyleTest was a case, all test use CodingStyleTest class with an
environment variable to control which business template is tested. In
this case, we want one log directory for each business template.

Introduce a getLogDirectoryPath method that's supposed to create and
return the path of the directory to use for log and override this method
for the special case of coding style test.