testnode/Upgrader: fix update when head is a merge commit
Updater was using this command to find the latest revision modifying a
path:

  git log -1 -- $PATH

but because of git log's history simplificiation [1] when the head of
the branch is a merge commit, the considered revision was not the merge
commit but its parent.

Use --full-history so that git does not simplify history and return the
merge commit if the latest commit is a merge.

[1]: https://git-scm.com/docs/git-log#_history_simplification

See also https://stackoverflow.com/a/50724301/7607763