erp5_full_text_mroonga_catalog: Use finer-grained serialization_tag.
The reason why root document is being used as serialization_tag is to cover
document which get recursively reindexed: such indexation activity would
not be able to be validated against all the the serializatio_tag of
documents it will process. But here, fulltext indexations are all spawned
per-document anyway, so we can use such per-document values.