web_renderjs_ui: update translation data in post-upgarde
Automate the process of running "Update Translation Data" in
post-upgarde step when the translation data has changed.

This means that translation scripts (such as gadget_translation_data.js
for default ERP5JS gadget) will be updated automatically after each
update and it becomes wrong to have multiple sites using the same
translation gadget for different translations.

If that happens, the upgrader tries to detect a loop and raise an
easy-to-understand error in that case