pdm: initialise domain tree on stock browser
To be useful, stock browser must show only "our" nodes, seeing all
suppliers with a negative stock or clients with a positive stock is not
interesting, so users always have to start by configuring their listbox
to domain tree node and select a site.

This was not really user friendly, but became a bigger problem with
ERP5JS listboxes that do not haave domain tree feature.

To improve this, initialize the selection with the preferred site
category.

Listbox does not support to initialize default values of domain tree
like, so we do this in a TALES expression for "selection_name", but
that's not really elegant.