pdm: repair converted inventory report listbox
getConvertedInventoryList is deprectated and use mergeResults which
does not support brain classes, so the brain.node_title was no longer
correct.
Use a simpler approach, same as Resource_getInventoryList ("stock
browser"), ie: getFutureInventoryList to get all future stocks +
getCurrentInventory and getAvailableInventory for each brain. It's a bit
less efficient because we now do 2 getInventory for each line instead of
3 getInventoryList merged, but that should be enough.

Also:
 - use float fields to format numbers properly
 - activate domain tree for site (instead of showing all nodes)